+++ src/main/java/com/globalshopper/GlobalShopper/repository/CommandeGroupeeRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface CommandeGroupeeRepository extends JpaRepository<CommandeGroupee, Long> {
+public interface CommandeGroupeeRepository extends JpaRepository<CommandeGroupee, Integer> {
     Optional<CommandeGroupee> findByProduitAndStatus(Produit produit, OrderStatus orderStatus);
 }
