+++ src/main/java/com/globalshopper/GlobalShopper/repository/ParticipationRepository.java
@@ -3,5 +3,5 @@ package com.globalshopper.GlobalShopper.repository;
 import com.globalshopper.GlobalShopper.entity.Participation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ParticipationRepository extends JpaRepository<Participation, Long> {
+public interface ParticipationRepository extends JpaRepository<Participation, Integer> {
 }
